--- app/lib/data.json
@@ -1,20 +1,0 @@
-{
-    "markdown": 1,
-    "code": 1,
-    "features": 1,
-    "next-js": 6,
-    "math": 1,
-    "ols": 1,
-    "github": 1,
-    "guide": 5,
-    "tailwind": 3,
-    "hello": 1,
-    "holiday": 1,
-    "canada": 1,
-    "images": 1,
-    "feature": 2,
-    "writings": 1,
-    "book": 1,
-    "reflection": 1,
-    "multi-author": 1
-}
