--- app/lib/data.ts
@@ -1,7 +1,0 @@
-export const headerList = [
-    { title: 'Home', path: '/' },
-    { title: 'Posts', path: '/posts' },
-    { title: 'Docs', path: '/docs' },
-    { title: 'Concat', path: '/concat' },
-    { title: 'About', path: '/about' }
-]
