--- app/page.tsx
@@ -1,43 +1,0 @@
-import Link from 'next/link'
-import type { Metadata } from 'next'
-
-export const metadata: Metadata = {
-  title: "Leafage",
-  description: "Generated by create next app",
-}
-
-export default function Home() {
-  return (
-    <section className="max-w-4xl mx-auto py-40">
-      <div className="hidden sm:mb-8 sm:flex sm:justify-center">
-        <div className="rounded-full py-1 px-3 text-sm leading-6 text-neutral-600 dark:text-neutral-400 ring-1 ring-neutral-900/10 hover:ring-neutral-900/20">
-          Announcing our next round of funding.{' '}
-          <Link href="/docs" className="font-semibold text-lime-600">
-            <span className="inset-0" aria-hidden="true" />
-            Read more <span aria-hidden="true">&rarr;</span>
-          </Link>
-        </div>
-      </div>
-      <div className="text-center">
-        <h1 className="text-4xl font-bold tracking-tight text-neutral-900 dark:text-neutral-300 sm:text-6xl">
-          Help to create your service security & easily.
-        </h1>
-        <p className="mt-6 text-lg leading-8 text-neutral-600 dark:text-neutral-400">
-          Anim aute id magna aliqua ad ad non deserunt sunt. Qui irure qui lorem cupidatat commodo. Elit sunt amet
-          fugiat veniam occaecat fugiat aliqua.
-        </p>
-        <div className="mt-10 flex items-center justify-center gap-x-6">
-          <a
-            href="#"
-            className="rounded-full bg-lime-600 px-4 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-lime-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-lime-600"
-          >
-            Get started
-          </a>
-          <a href="#" className="text-sm font-semibold leading-6 text-neutral-900 dark:text-neutral-300">
-            Learn more <span aria-hidden="true">→</span>
-          </a>
-        </div>
-      </div>
-    </section>
-  )
-}
