+++ package.json
@@ -9,7 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "clx": "^1.0.0",
     "next": "^15.1.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
@@ -24,6 +23,7 @@
     "unified": "^11.0.5"
   },
   "devDependencies": {
+    "@eslint/eslintrc": "^3.2.0",
     "@heroicons/react": "^2.1.5",
     "@tailwindcss/typography": "^0.5.15",
     "@types/node": "^20.12.7",
@@ -32,7 +32,8 @@
     "autoprefixer": "^10.4.20",
     "clsx": "^2.1.0",
     "eslint": "^9.14.0",
-    "eslint-config-next": "^15.0.2",
+    "eslint-config-next": "^15.1.0",
+    "gray-matter": "^4.0.3",
     "postcss": "^8.4.47",
     "tailwindcss": "^3.4.14",
     "typescript": "^5.6.3"
