+++ src/app/_components/header.tsx
@@ -5,8 +5,8 @@ import Link from 'next/link'
 import Image from 'next/image'
 import { clsx } from 'clsx'
 import { MagnifyingGlassIcon, Bars3Icon } from '@heroicons/react/24/outline'
-import ThemeSwitch from '@/app/ui/theme-switch'
-import { headerList } from '@/app/lib/data'
+import ThemeSwitch from '@/app/_components/theme-switch'
+import { headerList } from '@/lib/data'
 
 export default function Header() {
   const pathname = usePathname()
@@ -46,12 +46,12 @@ export default function Header() {
         </ul>
       </nav>
       <div className="hidden lg:inline-flex lg:w-1/4 items-center justify-end space-x-1">
-        <ThemeSwitch />
         <div className="relative">
           <button type="button" title='search_button' className="w-9 h-9 rounded-full hover:bg-lime-50 dark:hover:bg-neutral-800 focus:outline-none flex items-center justify-center transition-colors duration-200" id="dark" aria-expanded="false">
             <MagnifyingGlassIcon className="h-5 w-5 text-neutral-700 dark:text-neutral-300" aria-hidden="true" />
           </button>
         </div>
+        <ThemeSwitch />
         <div className="relative">
           <Link href="https://github.com/little3201" target="_blank" className="w-9 h-9 rounded-full text-neutral-700 dark:text-neutral-300 hover:bg-lime-50 dark:hover:bg-neutral-800 focus:outline-none flex items-center justify-center transition-colors duration-200" id="search" type="button" aria-expanded="false">
             <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="feather feather-github"><path d="M9 19c-5 1.5-5-2.5-7-3m14 6v-3.87a3.37 3.37 0 0 0-.94-2.61c3.14-.35 6.44-1.54 6.44-7A5.44 5.44 0 0 0 20 4.77 5.07 5.07 0 0 0 19.91 1S18.73.65 16 2.48a13.38 13.38 0 0 0-7 0C6.27.65 5.09 1 5.09 1A5.07 5.07 0 0 0 5 4.77a5.44 5.44 0 0 0-1.5 3.78c0 5.42 3.3 6.61 6.44 7A3.37 3.37 0 0 0 9 18.13V22"></path></svg>
