+++ src/app/page.tsx
@@ -1,0 +1,70 @@
+import Link from 'next/link'
+import type { Metadata } from 'next'
+
+export const metadata: Metadata = {
+  title: "Leafage",
+  description: "Generated by create next app",
+}
+
+export default function Index() {
+  return (
+    <div className='relative'>
+      <div className="absolute inset-x-0 -top-40 -z-10 transform-gpu overflow-hidden blur-3xl">
+        <svg
+          className="relative left-[calc(50%-11rem)] -z-10 h-[21.1875rem] max-w-none -translate-x-1/2 rotate-[30deg] sm:left-[calc(50%-30rem)] sm:h-[42.375rem]"
+          viewBox="0 0 1155 678"
+        >
+          <path
+            fill="url(#45de2b6b-92d5-4d68-a6a0-9b9b2abad533)"
+            fillOpacity=".3"
+            d="M317.219 518.975L203.852 678 0 438.341l317.219 80.634 204.172-286.402c1.307 132.337 45.083 346.658 209.733 145.248C936.936 126.058 882.053-94.234 1031.02 41.331c119.18 108.451 130.68 295.337 121.53 375.223L855 299l21.173 362.054-558.954-142.079z"
+          />
+          <defs>
+            <linearGradient
+              id="45de2b6b-92d5-4d68-a6a0-9b9b2abad533"
+              x1="1155.49"
+              x2="-78.208"
+              y1=".177"
+              y2="474.645"
+              gradientUnits="userSpaceOnUse"
+            >
+              <stop stopColor="#84cc16" />
+              <stop offset={1} stopColor="#15803d" />
+            </linearGradient>
+          </defs>
+        </svg>
+      </div>
+      <section className="max-w-4xl mx-auto my-40">
+        <div className="hidden sm:mb-8 sm:flex sm:justify-center">
+          <div className="rounded-full py-1 px-3 text-sm leading-6 text-neutral-600 dark:text-neutral-400 ring-1 ring-neutral-900/10 hover:ring-neutral-900/20">
+            Announcing our next round of funding.{' '}
+            <Link href="/docs" className="font-semibold text-lime-600">
+              <span className="inset-0" aria-hidden="true" />
+              Read more <span aria-hidden="true">&rarr;</span>
+            </Link>
+          </div>
+        </div>
+        <div className="text-center">
+          <h1 className="text-4xl font-bold tracking-tight text-neutral-900 dark:text-neutral-300 sm:text-6xl">
+            Help to create your service security & easily.
+          </h1>
+          <p className="mt-6 text-lg leading-8 text-neutral-600 dark:text-neutral-400">
+            Anim aute id magna aliqua ad ad non deserunt sunt. Qui irure qui lorem cupidatat commodo. Elit sunt amet
+            fugiat veniam occaecat fugiat aliqua.
+          </p>
+          <div className="mt-10 flex items-center justify-center gap-x-6">
+            <a
+              href="#"
+              className="rounded-full bg-lime-600 px-4 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-lime-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-lime-600"
+            >
+              Get started
+            </a>
+            <a href="#" className="text-sm font-semibold leading-6 text-neutral-900 dark:text-neutral-300">
+              Learn more <span aria-hidden="true">→</span>
+            </a>
+          </div>
+        </div>
+      </section>
+    </div>
+  )
+}
