+++ src/interfaces/doc.ts
@@ -1,0 +1,3 @@
+export type Doc = {
+  id: number
+}
