+++ src/interfaces/pagination.ts
@@ -1,0 +1,4 @@
+export type Pagination = {
+  totalPages: number
+  page: number
+}
