+++ src/interfaces/post.ts
@@ -1,0 +1,11 @@
+export type Post = {
+  slug: string
+  title: string;
+  date: string;
+  excerpt: string;
+  ogImage: {
+    url: string;
+  };
+  content: string;
+  tags: string[];
+}
