+++ src/lib/data.ts
@@ -1,0 +1,6 @@
+export const headerList = [
+  { title: 'Home', path: '/' },
+  { title: 'Posts', path: '/posts' },
+  { title: 'Docs', path: '/docs' },
+  { title: 'About', path: '/about' }
+]
