+++ src/lib/md-convert.ts
@@ -1,0 +1,24 @@
+import { unified } from 'unified'
+import remarkParse from 'remark-parse'
+import remarkGfm from 'remark-gfm'
+import remarkRehype from 'remark-rehype'
+import rehypeSlug from 'rehype-slug'
+import rehypeExternalLinks from 'rehype-external-links'
+import rehypeHighlight from 'rehype-highlight'
+import rehypeSanitize from 'rehype-sanitize'
+import rehypeStringify from 'rehype-stringify'
+
+
+export async function parseMarkdown(doc: string) {
+  const file = await unified()
+    .use(remarkParse, { fragment: true })
+    .use(remarkGfm)
+    .use(remarkRehype)
+    .use(rehypeSlug)
+    .use(rehypeSanitize)
+    .use(rehypeExternalLinks, { target: '_blank', rel: ['nofollow'] })
+    .use(rehypeHighlight)
+    .use(rehypeStringify)
+    .process(doc)
+  return String(file)
+}
